Adding a .NET Framework 4.8 (Windows) version
Also swapped out the AES encryption mode of CFB for CBC of the bearer token
encryption/decryption.  The CFB mode was throwing a padding issue on
Windows .NET 4.8 but not on .NET Core 7....

Also, the eBearer field now is formatted as:
[base64encoded aes iv].[base64-encoded aes-256-cbc ciphertext]

Bot the .NET Core 7 and the .NET 4.8 versions have been tested against
the DEV endpoint.